{{cookiecutter.project_slug}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.project_slug}}"
version = "0.1.0"
description = "{{cookiecutter.description}}"
authors = ["{{cookiecutter.full_name}} <{{cookiecutter.email}}>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.63.0"
pydantic = "^1.7.3"
uvicorn = "^0.13.3"
python-json-logger = "^2.0.1"
sentry-sdk = "^0.19.5"
sqlalchemy = "^1.3.23"
psycopg2 = "^2.8.6"
httpx = "^0.16.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
pytest-cov = "^2.11.1"
requests = "^2.25.1"

[build-system]
requires = ["poetry>=1.0.10"]
build-backend = "poetry.masonry.api"
